requirements.txt
pyyaml==5.1
torch==2.2.0
torchvision==0.17.0
git+https://github.com/facebookresearch/detectron2.git
opencv-python
google~=3.0.0


jupyter~=1.0.0
pip~=24.0
attrs~=23.2.0
pillow~=10.4.0
tornado~=6.4
Jinja2~=3.1.3
filelock~=3.15.4
typing_extensions~=4.12.2
packaging~=23.2
platformdirs~=4.2.0
pandas~=2.2.0
numpy~=1.26.4